catalog: support nested anyOf/allOf filters in NextEntitiesCatalog
In order to allow individual authorization filters to return complex
filter queries, the catalog needs to support nested anyOf/allOf objects
in filters. This work is incomplete - it's missing unit tests, and
there's also a todo to fix a bug uncovered along the way, that we should
at least open an issue for, if not fix.

Signed-off-by: Mike Lewis <mtlewis@users.noreply.github.com>
Co-authored-by: Natasha Issayeva <nissayeva@users.noreply.github.com>